[codex] Finish autoresearch report backlog cleanup
Merge PR #148 after local npm run check, git diff --check, and GitHub CI passed.
