+++ plugins/codex-autoresearch/docs/maintainers.md
@@ -82,12 +82,13 @@ Before publishing, inspect the package artifact itself. Use dry-run pack output
 for routine review, then create and extract a real tarball for release smoke.
 The shipped `scripts/*.mjs` shims depend on `dist/`, but `dist/` is generated
 and ignored in the Git tree. If a Git marketplace source checkout is missing
-`dist/`, the CLI launcher downloads the matching GitHub release tarball plus
-`codex-autoresearch-<version>.tgz.sha256`, verifies the SHA-256 entry names that
-exact tarball, verifies the packaged name/version, and only then extracts
-`dist/` into the plugin cache before importing the runtime. A publishable
-release tarball must include the built runtime, publish the adjacent checksum
-asset, exclude authored source and tests, ship no MCP launcher/config, and pass
+`dist/`, the CLI launcher calls `scripts/bootstrap-runtime.mjs` to download the
+matching GitHub release tarball plus
+`codex-autoresearch-<version>.tgz.sha256`, verify the SHA-256 entry names that
+exact tarball, verify the packaged name/version, and only then extract `dist/`
+into the plugin cache before importing the runtime. A publishable release
+tarball must include the built runtime, publish the adjacent checksum asset,
+exclude authored source and tests, ship no MCP launcher/config, and pass
 `node <extracted-package>/scripts/autoresearch.mjs --help`.
 
 Do not push release tags by hand. After a synchronized version bump lands on `main`, the `Auto Release` GitHub Actions workflow compares the previous and current package versions and calls the reusable `Release` workflow when the package version changed. The release workflow still runs the checks, builds and smoke-tests the tarball, refuses pre-existing tags, and only then creates the GitHub release/tag with the tarball asset attached. Use manual `Release` dispatch only as an explicit recovery path with the package version. This keeps update clients on the previous release until the new install artifact exists.
